fix: resolve post-merge compilation errors and add test infrastructure
- Fix std::fabsf → std::fabs in IntroAnimation.cpp (not in std namespace)
- Add insolation_flux field to ExoplanetData (referenced by blackhole-v UI)
- Fix buildFullRenderParamsPrompt → buildRenderParamsPrompt in InferenceEngine
- Use system libcurl (apt: libcurl4-openssl-dev) instead of FetchContent
- Add Catch2 + test subdirectory back to CMakeLists.txt
- Add pugixml to Dependencies.cmake for OEC XML parsing
- Build: 100% clean, zero warnings, 516 test assertions passing

+++ src/intro/IntroAnimation.cpp
@@ -386,7 +386,7 @@ float IntroAnimation::getUIAlpha() const {
 void IntroAnimation::syncBorderToWindow(float x, float y, float w, float h) {
     const float dx = x - m_borderX;
     const float dy = y - m_borderY;
-    if (std::fabsf(dx) < 0.5f && std::fabsf(dy) < 0.5f) return; // already aligned
+    if (std::fabs(dx) < 0.5f && std::fabs(dy) < 0.5f) return; // already aligned
 
     for (auto& p : m_particles) {
         if (!p.goToBorder) continue;
